test: 🔄 move DeepWritable to a separate file
